==> spatial_HYP/code/04-Feature Selection, Harmony, MNN/01-Feature selection.Rmd <==
---
title: "01-Feature selection"
author: "Bernard Mulvey"
date: "2023-11-16"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different:
knitr::opts_knit$set(root.dir = here::here())

library(ggplot2)
library(data.table)
library(gridExtra)
require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")
library(SpatialExperiment)
library(scater)
library(nnSVG)
library(BiocParallel)
library(scran)

# ggplot defaults:
theme_set(theme_bw()+theme(axis.text.x = element_text(size = 14), axis.title.x = element_text(size = 16), axis.text.y = element_text(size = 14), axis.title.y = element_text(size =16), plot.title = element_text(size = 20,hjust=0.5), strip.text = element_text(size=18), legend.text = element_text(size=10), legend.title = element_text(size=11,hjust=0.5)))
```


# Load pre-processed SPE and demographics; attach to colData
```{r}
spesessionloader <- function(){
  hyp2 <- readRDS("processed-data/03-QC_filters/hypN10_umi210_gene126_chrm50_spotsweeped_lognorm_rotsNmirrors_071924.RDS")
  
  # make sure colnames of the spe are unique (i.e., colData(hyp2)$key)
  colnames(hyp2) <- hyp2$key
  
  ## for efficiency during these steps, drop the imgData slot
  hyp2@int_metadata$imgData <- DataFrame()
  gc(full=T)
  
  return(hyp2)
}

hyp2 <- spesessionloader()
```

### mean-variance relationships: blocking on sample_id (not donor)
```{r}
hyp2.vars <- modelGeneVar(hyp2,block=hyp2$sample_id)

varperblock <- hyp2.vars$per.block

# examine per-sample (block) variances by plotting each mean-variance curve and storing in a list to arrange later as a single output
plts <- list()

i <- 1
for (i in c(1:length(varperblock))){
    hyp2.vars.dt <- cbind(as.data.table(as.character(rownames(varperblock[[i]]))),varperblock[[i]]$total)
    hyp2.means.dt <- cbind(as.data.table(as.character(rownames(varperblock[[i]]))),varperblock[[i]]$mean)

    setnames(hyp2.vars.dt,2,"var_log_xpr")
    setnames(hyp2.means.dt,2,"mean_log_xpr")
    hyp2.mnvar.pltdat <- merge(hyp2.means.dt,hyp2.vars.dt,by="V1")
    plts[[i]] <- ggplot(hyp2.mnvar.pltdat, aes(x = mean_log_xpr, y = var_log_xpr)) +
        geom_point() +
        geom_function(fun=metadata(varperblock[[i]])$trend,col="blue")+
        ggtitle(paste0("Mean-var curve-",names(varperblock)[i],"\n sample_id blocking"))
    rm(hyp2.mnvar.pltdat,hyp2.vars.dt,hyp2.means.dt)
}

## arrange the per sample plots generated above into a single output
pdf("plots/04-Feature Selection, PCA, Harmony/01-samplewiseblocking_meanvars.pdf",height=3*length(unique(hyp2$sample_id))+2,width=10)
do.call("grid.arrange",c(plts,ncol=2))
dev.off()

## remove clutter from workspace
rm(plts,varperblock,i)
```

# get top 10, 20 %ile HVGs from sample_id-blocking'd variances
```{r}
hyp2.hvg10 <- scran::getTopHVGs(hyp2.vars,prop = 0.1)
hyp2.hvg20 <- scran::getTopHVGs(hyp2.vars,prop=0.2)

rm(hyp2.vars)
```

#### run nnSVG: does not take batch variables nor does it take in reduced dims. run samplewise.

### the workhorse function nnSVG can use parallel processing. the parallelization is within the nnsvg call, so no need to match cpu # to sample #
```{r}
# SUPER CRITICAL IMPORTANT NOTES: 
# 1. RECALCULATE LOG COUNTS AFTER FILTERING ^
# 2.nnSVG works ONE SAMPLE (one capture area) at a time. So need to wrap this in a loop. 
# 3. as such, perform filtering PER SAMPLE. 
# filter_genes_ncounts (N) and filter_genes_pcspots (S) work together: genes will be filtered to those with at least N counts in at least S% of spots.  filter_genes_pcspots expects a %age as 0-100, NOT a fraction.
# tried N=2 (instead of 3) and S = 0.25 instead of 0.5, since most spots are pretty diverse and high-depth (and the cell type diversity is also unknown), but didn't get past the first sample in 3 hours on 10 CPUs. so now trying 3, 0.5 .

caps <- unique(hyp2$sample_id)
hypsamps <- lapply(caps,FUN=function(x){hyp2[,hyp2$sample_id==x]})
names(hypsamps) <- caps

### clean up for the big run
rm(hyp2,caps)
gc(full=T)

## make list to hold results
res_list <- sapply(names(hypsamps),FUN=function(x){NULL})
```


Iterate nnsvg by sample
```{r}
i<-1
for (i in c(1:length(hypsamps))){
    #setup - swap the sample out of the list. this progressively clears space as we get through samples, SAVING memory as we go. 
  hyp2.samp <- hypsamps[[i]]
  hypsamps[[i]] <- "started"
  
  ### to identify new rowData columns added by nnsvg, we need to know how many there are to start with:
  rdind <- ncol(rowData(hyp2.samp))
  
  message(paste0(Sys.time()," - Starting sample ",names(hypsamps)[i]," (",i," of ",length(hypsamps),")"))
  
  hyp2.samp <- filter_genes(hyp2.samp,filter_genes_ncounts = 3,filter_genes_pcspots = 0.5,filter_mito = F)
    
  #recalculate logcounts after filtering for nnsvg-able genes
  hyp2.samp <- computeLibraryFactors(hyp2.samp)
  hyp2.samp <- logNormCounts(hyp2.samp)
    
  # main call -- defaults to logcounts but shown for posterity; X if using covariates; n_neighbors defaults to 10, slower but potentially more granular with higher values eg 15, order = don't touch unless you have < 70 spots total; n_threads or BPPARAM for parallelization
  hyp2.samp <- nnSVG(hyp2.samp,assay_name = "logcounts",n_threads = 10)
  
    # extract results; using index of the last column of original rowdata to determine where nnSVG-appended columns begin; add nnSVG_ to those column names so we can tell them apart
  rdstartind <- rdind+1
  colnames(rowData(hyp2.samp))[rdstartind:ncol(rowData(hyp2.samp))] <- paste0("nnSVG_",colnames(rowData(hyp2.samp))[rdstartind:ncol(rowData(hyp2.samp))])

  # extract the nnsvg results from the rowData and store it in the result list. we don't need the rownames for storage since they're redundant with the ensembl identifier column in rowdata.
  outtab <- as.data.table(as.data.frame(rowData(hyp2.samp)),keep.rownames=T)
  res_list[[i]] <- copy(outtab)
  names(res_list)[i] <- names(hypsamps)[i]
  
  rm(outtab,rdstartind,hyp2.samp,rdind)
  gc(full=T)
}

saveRDS(res_list,"processed-data/04-feature_selection/01-hyp2n10_nnSVG_samplewise_n3s0.5.RDS")

## make an environment copy of the nnsvg results with a more descriptive name we can come back to
nnsvg_N3Spt5 <- res_list

## further cleanup
rm(i,caps,res_list,hypsamps)
gc(full=T)
```

### get the mean nnSVG rank for each gene among samples where the gene was nominally significant for spatial variability.
```{r}
## reload main SPE
hyp2 <- spesessionloader()

i<-1
for (i in c(1:sum(unique(colData(hyp2)$sample_id)%in%names(nnsvg_N3Spt5)))){
    if(i==1){
        nnsvg_ranks <- nnsvg_N3Spt5[[i]][nnSVG_padj<0.05]
        nnsvg_ranks <- nnsvg_ranks[,.(gene_id,gene_name,nnSVG_rank)]
        setnames(nnsvg_ranks,"nnSVG_rank",names(nnsvg_N3Spt5)[i])
    }
    else{
        tmp <- nnsvg_N3Spt5[[i]][nnSVG_padj<0.05]
        tmp <- tmp[,.(gene_id,gene_name,nnSVG_rank)]
        setnames(tmp,"nnSVG_rank",names(nnsvg_N3Spt5)[i])
        nnsvg_ranks <- merge.data.table(nnsvg_ranks, tmp, by=c("gene_id","gene_name"), all.x=T, all.y=T, incomparables = NA)
        rm(tmp)
    }
}
rm(i)

nnsvg_ranks[,meanrank:=rowMeans(.SD,na.rm=T),.SDcols=c(3:ncol(nnsvg_ranks))]
nnsvg_ranks <- nnsvg_ranks[,samples_signif:=rowSums(!is.na(.SD)),.SDcols=c(3:ncol(nnsvg_ranks))]

fwrite(nnsvg_ranks,"processed-data/04-feature_selection/01-hyp2n10_meangeneranks_n10_nnSVG_samplewise_n3s0.5.txt",sep='\t',quote=F)
```

### retrieve top M or N nnSVG genes (where M = number of genes in the HVG 10%ile set and N = number in the top 20%ile HVG set); save each vector of variable genes.
```{r}
nnsvg_ranks <- nnsvg_ranks[samples_signif>2] # at least 3 samples with nominally significant nnSVG results (meaning at least 2 donors w " " " " " ") 
setorderv(nnsvg_ranks,"meanrank",1) # ascending order
nnsvg10 <- nnsvg_ranks[c(1:length(hyp2.hvg10)),gene_id]
nnsvg20 <- nnsvg_ranks[c(1:length(hyp2.hvg20)),gene_id]

featuresets <- list(hyp2.hvg10=hyp2.hvg10,hyp2.hvg20=hyp2.hvg20,nnsvg10=nnsvg10,nnsvg20=nnsvg20)

saveRDS(featuresets,"processed-data/04-feature_selection/01-hyp2n10_HVG10_20pctile_sample-blocked_nnSVG-mnrank-in-nomsigsamps_top10_20pctile.RDS")
```

# session info follows call #
```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] scran_1.32.0                BiocParallel_1.38.0        
 [3] nnSVG_1.8.0                 scater_1.32.0              
 [5] scuttle_1.14.0              SpatialExperiment_1.14.0   
 [7] SingleCellExperiment_1.26.0 SummarizedExperiment_1.34.0
 [9] Biobase_2.64.0              GenomicRanges_1.56.1       
[11] GenomeInfoDb_1.40.1         IRanges_2.38.1             
[13] S4Vectors_0.42.1            BiocGenerics_0.50.0        
[15] MatrixGenerics_1.16.0       matrixStats_1.3.0          
[17] colorout_1.3-0.2            gridExtra_2.3              
[19] data.table_1.15.4           ggplot2_3.5.1              
[21] rlang_1.1.4                

loaded via a namespace (and not attached):
 [1] pbapply_1.7-2             magrittr_2.0.3            compiler_4.4.1           
 [4] DelayedMatrixStats_1.26.0 vctrs_0.6.5               pkgconfig_2.0.3          
 [7] crayon_1.5.3              rdist_0.0.5               fastmap_1.2.0            
[10] magick_2.8.3              XVector_0.44.0            labeling_0.4.3           
[13] utf8_1.2.4                BRISC_1.0.5               rmarkdown_2.27           
[16] UCSC.utils_1.0.0          ggbeeswarm_0.7.2          xfun_0.45                
[19] bluster_1.14.0            zlibbioc_1.50.0           beachmat_2.20.0          
[22] jsonlite_1.8.8            DelayedArray_0.30.1       irlba_2.3.5.1            
[25] parallel_4.4.1            cluster_2.1.6             R6_2.5.1                 
[28] limma_3.60.3              Rcpp_1.0.13               knitr_1.48               
[31] Matrix_1.7-0              igraph_2.0.3              tidyselect_1.2.1         
[34] rstudioapi_0.16.0         abind_1.4-5               yaml_2.3.9               
[37] viridis_0.6.5             codetools_0.2-20          lattice_0.22-6           
[40] tibble_3.2.1              withr_3.0.0               evaluate_0.24.0          
[43] pillar_1.9.0              generics_0.1.3            rprojroot_2.0.4          
[46] sparseMatrixStats_1.16.0  munsell_0.5.1             scales_1.3.0             
[49] glue_1.7.0                metapod_1.12.0            tools_4.4.1              
[52] BiocNeighbors_1.22.0      ScaledMatrix_1.12.0       locfit_1.5-9.10          
[55] RANN_2.6.1                grid_4.4.1                edgeR_4.2.0              
[58] colorspace_2.1-0          GenomeInfoDbData_1.2.12   beeswarm_0.4.0           
[61] BiocSingular_1.20.0       vipor_0.4.7               cli_3.6.3                
[64] rsvd_1.0.5                fansi_1.0.6               S4Arrays_1.4.1           
[67] viridisLite_0.4.2         dplyr_1.1.4               gtable_0.3.5             
[70] digest_0.6.36             SparseArray_1.4.8         ggrepel_0.9.5            
[73] dqrng_0.4.1               farver_2.1.2              rjson_0.2.21             
[76] htmltools_0.5.8.1         lifecycle_1.0.4           httr_1.4.7               
[79] here_1.0.1                statmod_1.5.0            
> sessioninfo::session_info()
─ Session info ──────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-07-22
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ──────────────────────────────────────────────────────────────────────
 ! package              * version  date (UTC) lib source
   abind                  1.4-5    2016-07-21 [1] CRAN (R 4.4.0)
   beachmat               2.20.0   2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   beeswarm               0.4.0    2021-06-01 [1] CRAN (R 4.4.0)
   Biobase              * 2.64.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocGenerics         * 0.50.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocNeighbors          1.22.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocParallel         * 1.38.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   BiocSingular           1.20.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   bluster                1.14.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BRISC                  1.0.5    2022-04-29 [1] CRAN (R 4.4.0)
 P cli                    3.6.3    2024-06-21 [2] CRAN (R 4.4.0)
   cluster                2.1.6    2023-12-01 [1] CRAN (R 4.4.1)
   codetools              0.2-20   2024-03-31 [1] CRAN (R 4.4.1)
   colorout             * 1.3-0.2  2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   colorspace             2.1-0    2023-01-23 [1] CRAN (R 4.4.0)
   crayon                 1.5.3    2024-06-20 [1] CRAN (R 4.4.0)
   data.table           * 1.15.4   2024-03-30 [2] CRAN (R 4.4.0)
   DelayedArray           0.30.1   2024-05-07 [1] Bioconductor 3.19 (R 4.4.0)
   DelayedMatrixStats     1.26.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   digest                 0.6.36   2024-06-23 [1] CRAN (R 4.4.0)
   dplyr                  1.1.4    2023-11-17 [1] CRAN (R 4.4.0)
   dqrng                  0.4.1    2024-05-28 [1] CRAN (R 4.4.0)
   edgeR                  4.2.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   evaluate               0.24.0   2024-06-10 [1] CRAN (R 4.4.0)
   fansi                  1.0.6    2023-12-08 [1] CRAN (R 4.4.0)
   farver                 2.1.2    2024-05-13 [1] CRAN (R 4.4.0)
   fastmap                1.2.0    2024-05-15 [1] CRAN (R 4.4.0)
   generics               0.1.3    2022-07-05 [1] CRAN (R 4.4.0)
   GenomeInfoDb         * 1.40.1   2024-05-24 [1] Bioconductor 3.19 (R 4.4.0)
   GenomeInfoDbData       1.2.12   2024-05-01 [1] Bioconductor
   GenomicRanges        * 1.56.1   2024-06-12 [1] Bioconductor 3.19 (R 4.4.1)
   ggbeeswarm             0.7.2    2023-04-29 [1] CRAN (R 4.4.0)
   ggplot2              * 3.5.1    2024-04-23 [1] CRAN (R 4.4.0)
   ggrepel                0.9.5    2024-01-10 [1] CRAN (R 4.4.0)
   glue                   1.7.0    2024-01-09 [1] CRAN (R 4.4.0)
   gridExtra            * 2.3      2017-09-09 [1] CRAN (R 4.4.0)
   gtable                 0.3.5    2024-04-22 [1] CRAN (R 4.4.0)
   here                   1.0.1    2020-12-13 [1] CRAN (R 4.4.0)
   htmltools              0.5.8.1  2024-04-04 [1] CRAN (R 4.4.0)
   httr                   1.4.7    2023-08-15 [1] CRAN (R 4.4.0)
   igraph                 2.0.3    2024-03-13 [1] CRAN (R 4.4.0)
   IRanges              * 2.38.1   2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   irlba                  2.3.5.1  2022-10-03 [1] CRAN (R 4.4.0)
   jsonlite               1.8.8    2023-12-04 [1] CRAN (R 4.4.0)
   knitr                  1.48     2024-07-07 [1] CRAN (R 4.4.1)
   labeling               0.4.3    2023-08-29 [1] CRAN (R 4.4.0)
   lattice                0.22-6   2024-03-20 [1] CRAN (R 4.4.1)
   lifecycle              1.0.4    2023-11-07 [1] CRAN (R 4.4.0)
   limma                  3.60.3   2024-06-16 [1] Bioconductor 3.19 (R 4.4.0)
   locfit                 1.5-9.10 2024-06-24 [1] CRAN (R 4.4.0)
   magick                 2.8.3    2024-02-18 [1] CRAN (R 4.4.0)
   magrittr               2.0.3    2022-03-30 [1] CRAN (R 4.4.0)
   Matrix                 1.7-0    2024-04-26 [1] CRAN (R 4.4.1)
   MatrixGenerics       * 1.16.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   matrixStats          * 1.3.0    2024-04-11 [1] CRAN (R 4.4.0)
   metapod                1.12.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   munsell                0.5.1    2024-04-01 [1] CRAN (R 4.4.0)
   nnSVG                * 1.8.0    2024-05-16 [1] Bioconductor 3.19 (R 4.4.0)
   pbapply                1.7-2    2023-06-27 [1] CRAN (R 4.4.0)
   pillar                 1.9.0    2023-03-22 [1] CRAN (R 4.4.0)
   pkgconfig              2.0.3    2019-09-22 [1] CRAN (R 4.4.0)
   R6                     2.5.1    2021-08-19 [1] CRAN (R 4.4.0)
   RANN                   2.6.1    2019-01-08 [1] CRAN (R 4.4.0)
   Rcpp                   1.0.13   2024-07-17 [1] CRAN (R 4.4.1)
   rdist                  0.0.5    2020-05-04 [1] CRAN (R 4.4.0)
   rjson                  0.2.21   2022-01-09 [1] CRAN (R 4.4.0)
 P rlang                * 1.1.4    2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown              2.27     2024-05-17 [1] CRAN (R 4.4.0)
   rprojroot              2.0.4    2023-11-05 [1] CRAN (R 4.4.0)
   rstudioapi             0.16.0   2024-03-24 [1] CRAN (R 4.4.0)
   rsvd                   1.0.5    2021-04-16 [1] CRAN (R 4.4.0)
   S4Arrays               1.4.1    2024-05-20 [1] Bioconductor 3.19 (R 4.4.0)
   S4Vectors            * 0.42.1   2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   ScaledMatrix           1.12.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scales                 1.3.0    2023-11-28 [1] CRAN (R 4.4.0)
   scater               * 1.32.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scran                * 1.32.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scuttle              * 1.14.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   sessioninfo            1.2.2    2021-12-06 [1] CRAN (R 4.4.0)
   SingleCellExperiment * 1.26.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   SparseArray            1.4.8    2024-05-30 [1] Bioconductor 3.19 (R 4.4.0)
   sparseMatrixStats      1.16.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   SpatialExperiment    * 1.14.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   statmod                1.5.0    2023-01-06 [1] CRAN (R 4.4.0)
   SummarizedExperiment * 1.34.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   tibble                 3.2.1    2023-03-20 [1] CRAN (R 4.4.0)
   tidyselect             1.2.1    2024-03-11 [1] CRAN (R 4.4.0)
   UCSC.utils             1.0.0    2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   utf8                   1.2.4    2023-10-22 [1] CRAN (R 4.4.0)
   vctrs                  0.6.5    2023-12-01 [1] CRAN (R 4.4.0)
   vipor                  0.4.7    2023-12-18 [1] CRAN (R 4.4.0)
   viridis                0.6.5    2024-01-29 [1] CRAN (R 4.4.0)
   viridisLite            0.4.2    2023-05-02 [1] CRAN (R 4.4.0)
   withr                  3.0.0    2024-01-16 [1] CRAN (R 4.4.0)
   xfun                   0.45     2024-06-16 [1] CRAN (R 4.4.0)
   XVector                0.44.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   yaml                   2.3.9    2024-07-05 [1] CRAN (R 4.4.0)
   zlibbioc               1.50.0   2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

─────────────────────────────────────────────────────────────────────────────────
