==> spatial_HYP/code/04-Feature Selection, Harmony, MNN/02-PCA, Harmony and MNN.Rmd <==
---
title: "02-Harmony and MNN"
author: "Bernard Mulvey"
date: "2024-07-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(Biostrings)
library(ggplot2)
library(gridExtra)
library(SpatialExperiment)
library(spatialLIBD)
require(colorout)
library(batchelor)
library(harmony)
library(BiocParallel)

ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)

theme_set(theme_bw()+theme(axis.text.x = element_text(size = 14), axis.title.x = element_text(size = 16), axis.text.y = element_text(size = 14), axis.title.y = element_text(size =16), plot.title = element_text(size = 20,hjust=0.5), strip.text = element_text(size=18), legend.text = element_text(size=10), legend.title = element_text(size=11,hjust=0.5)))
```

### spe loadup. we alreaady calculated the log-normalized counts after QC and rotating/mirroring the samples for consistent anatomic orientation
```{r}
spesessionloader <- function(){
  hyp2 <- readRDS("processed-data/03-QC_filters/hypN10_umi210_gene126_chrm50_spotsweeped_lognorm_rotsNmirrors_072224.RDS")
  
  # make sure colnames of the spe are unique (i.e., colData(hyp2)$key)
  colnames(hyp2) <- hyp2$key
  
  ## for efficiency during these steps, drop the imgData slot
  hyp2@int_metadata$imgData <- DataFrame()
  gc(full=T)
  
  return(hyp2)
}

hyp2 <- spesessionloader()
```

### load feature sets
```{r}
feats <- readRDS("processed-data/04-feature_selection/01-hyp2n10_HVG10_20pctile_sample-blocked_nnSVG-mnrank-in-nomsigsamps_top10_20pctile.RDS")
names(feats) <- c("hvg10","hvg20","nnsvg10","nnsvg20")
```

### Dimensionality reductions with each feature set: standard PCA, standard UMAP; PCA/UMAP with harmony (defaults); PCA/UMAP with harmony (lambda=NULL); PCA/UMAP with MNN

### for harmony set the initial lambda to NULL such that harmony estimates the input value to prevent overcorrection.
```{r}
# set up BiocParallel for mnn
sbp <- MulticoreParam(8)

i<-1
for (i in c(1:length(feats))){
    # standard PCAs
    set.seed(42)
    hyp2 <- scater::runPCA(hyp2,subset_row=feats[[i]],name=paste0("PCA_",names(feats)[i]))

    # standard UMAPs (which require PCA first)
    set.seed(42)
    hyp2 <- scater::runUMAP(hyp2,dimred=paste0("PCA_",names(feats)[i]),name=paste0("UMAP_",names(feats)[i]))
    
    # harmony runs -- we need to temporarily rename the PCA for the feature set to "PCA" for runharmony to accept it
    reducedDimNames(hyp2)[which(reducedDimNames(hyp2)==paste0("PCA_",names(feats)[i]))] <- "PCA"
    
    # vanilla harmony (lambda not specified --> default value of 1)
    set.seed(42)
    hyp2 <- RunHarmony(hyp2,group.by.vars = "sample_id",reduction.save=paste0("HARMONYdflt_",names(feats)[i]))
    
    # data-based initialization of lambda (development ver of Harmony; requires explicit lambda=NULL)
    set.seed(42)
    hyp2 <- RunHarmony(hyp2,lambda=NULL,group.by.vars="sample_id",reduction.save=paste0("HARMONYlmbna_",names(feats)[i]))
    
    # restore the feature-set-specific name for the uncorrected PCA
    reducedDimNames(hyp2)[which(reducedDimNames(hyp2)=="PCA")] <- paste0("PCA_",names(feats)[i])
    
    # MNN by donor with k=30
    set.seed(42)
    mnnout <- reducedMNN(reducedDim(hyp2,paste0("PCA_",names(feats)[i])),batch = hyp2$brnum,k=30,BPPARAM=sbp)
    
    # extract reduced dimension table and append to SPE
    reducedDim(hyp2,paste0("mnn30_",names(feats)[i])) <- mnnout$corrected
    # remove intermediate object
    rm(mnnout)
    
    # UMAP of vanilla harmony
    set.seed(42)
    hyp2 <- scater::runUMAP(hyp2,dimred=paste0("HARMONYdflt_",names(feats)[i]),name=paste0("HARMONYdflt_",names(feats)[i],"_UMAP"))
    
    # UMAP of data-based initialization of lambda
    set.seed(42)
    hyp2 <- scater::runUMAP(hyp2,dimred=paste0("HARMONYlmbna_",names(feats)[i]),name=paste0("HARMONYlmbna_",names(feats)[i],"_UMAP"))
    
    # UMAP of MNN
    set.seed(42)
    hyp2 <- scater::runUMAP(hyp2,dimred=paste0("mnn30_",names(feats)[i]),name=paste0("mnn30_",names(feats)[i],"_UMAP"))
    
}

## check that we have 8*length(feats) reduced dims on the object
stopifnot(length(reducedDimNames(hyp2))==8*length(feats))

```

### save RDS of all the dimreds ####
```{r}
dimredout <- reducedDims(hyp2)

saveRDS(dimredout,"processed-data/04-feature_selection/02-hypfiltered_4featureset-pca-umap-harmonydefault-harmonylambdanull-mnn30_reducedDim_slot.RDS")
```

### reproducibility information
```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] BiocParallel_1.38.0         harmony_1.2.0              
 [3] Rcpp_1.0.13                 default_1.0.0              
 [5] batchelor_1.20.0            colorout_1.3-0.2           
 [7] spatialLIBD_1.16.2          SpatialExperiment_1.14.0   
 [9] SingleCellExperiment_1.26.0 SummarizedExperiment_1.34.0
[11] Biobase_2.64.0              GenomicRanges_1.56.1       
[13] MatrixGenerics_1.16.0       matrixStats_1.3.0          
[15] gridExtra_2.3               ggplot2_3.5.1              
[17] Biostrings_2.72.1           GenomeInfoDb_1.40.1        
[19] XVector_0.44.0              IRanges_2.38.1             
[21] S4Vectors_0.42.1            BiocGenerics_0.50.0        
[23] data.table_1.15.4           rlang_1.1.4                

loaded via a namespace (and not attached):
  [1] RcppAnnoy_0.0.22          later_1.3.2               BiocIO_1.14.0            
  [4] bitops_1.0-7              filelock_1.0.3            fields_16.2              
  [7] tibble_3.2.1              XML_3.99-0.17             lifecycle_1.0.4          
 [10] edgeR_4.2.0               doParallel_1.0.17         rprojroot_2.0.4          
 [13] lattice_0.22-6            magrittr_2.0.3            limma_3.60.3             
 [16] plotly_4.10.4             sass_0.4.9                rmarkdown_2.27           
 [19] jquerylib_0.1.4           yaml_2.3.9                httpuv_1.6.15            
 [22] spam_2.10-0               sessioninfo_1.2.2         cowplot_1.1.3            
 [25] DBI_1.2.3                 RColorBrewer_1.1-3        ResidualMatrix_1.14.1    
 [28] golem_0.4.1               maps_3.4.2                abind_1.4-5              
 [31] zlibbioc_1.50.0           purrr_1.0.2               RCurl_1.98-1.14          
 [34] rappdirs_0.3.3            GenomeInfoDbData_1.2.12   ggrepel_0.9.5            
 [37] irlba_2.3.5.1             DelayedMatrixStats_1.26.0 codetools_0.2-20         
 [40] DelayedArray_0.30.1       DT_0.33                   scuttle_1.14.0           
 [43] tidyselect_1.2.1          UCSC.utils_1.0.0          ScaledMatrix_1.12.0      
 [46] viridis_0.6.5             shinyWidgets_0.8.6        BiocFileCache_2.12.0     
 [49] GenomicAlignments_1.40.0  jsonlite_1.8.8            BiocNeighbors_1.22.0     
 [52] scater_1.32.0             iterators_1.0.14          foreach_1.5.2            
 [55] tools_4.4.1               glue_1.7.0                SparseArray_1.4.8        
 [58] xfun_0.45                 here_1.0.1                dplyr_1.1.4              
 [61] withr_3.0.0               BiocManager_1.30.23       fastmap_1.2.0            
 [64] fansi_1.0.6               digest_0.6.36             rsvd_1.0.5               
 [67] R6_2.5.1                  mime_0.12                 colorspace_2.1-0         
 [70] RSQLite_2.3.7             RhpcBLASctl_0.23-42       config_0.3.2             
 [73] utf8_1.2.4                tidyr_1.3.1               generics_0.1.3           
 [76] rtracklayer_1.64.0        httr_1.4.7                htmlwidgets_1.6.4        
 [79] S4Arrays_1.4.1            uwot_0.2.2                pkgconfig_2.0.3          
 [82] gtable_0.3.5              blob_1.2.4                htmltools_0.5.8.1        
 [85] dotCall64_1.1-1           scales_1.3.0              png_0.1-8                
 [88] attempt_0.3.1             knitr_1.48                rstudioapi_0.16.0        
 [91] rjson_0.2.21              curl_5.2.1                cachem_1.1.0             
 [94] BiocVersion_3.19.1        parallel_4.4.1            vipor_0.4.7              
 [97] AnnotationDbi_1.66.0      restfulr_0.0.15           pillar_1.9.0             
[100] grid_4.4.1                vctrs_0.6.5               promises_1.3.0           
[103] BiocSingular_1.20.0       dbplyr_2.5.0              beachmat_2.20.0          
[106] xtable_1.8-4              beeswarm_0.4.0            paletteer_1.6.0          
[109] evaluate_0.24.0           magick_2.8.3              cli_3.6.3                
[112] locfit_1.5-9.10           compiler_4.4.1            Rsamtools_2.20.0         
[115] crayon_1.5.3              rematch2_2.1.2            ggbeeswarm_0.7.2         
[118] viridisLite_0.4.2         munsell_0.5.1             lazyeval_0.2.2           
[121] Matrix_1.7-0              ExperimentHub_2.12.0      benchmarkme_1.0.8        
[124] sparseMatrixStats_1.16.0  bit64_4.0.5               KEGGREST_1.44.1          
[127] statmod_1.5.0             shiny_1.8.1.1             AnnotationHub_3.12.0     
[130] igraph_2.0.3              memoise_2.0.1             bslib_0.7.0              
[133] benchmarkmeData_1.0.4     bit_4.0.5                
> sessioninfo::session_info()
─ Session info ──────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-07-23
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ──────────────────────────────────────────────────────────────────────
 ! package              * version   date (UTC) lib source
   abind                  1.4-5     2016-07-21 [1] CRAN (R 4.4.0)
   AnnotationDbi          1.66.0    2024-05-01 [1] Bioconductor 3.19 (R 4.4.0)
   AnnotationHub          3.12.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   attempt                0.3.1     2020-05-03 [1] CRAN (R 4.4.0)
   batchelor            * 1.20.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   beachmat               2.20.0    2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   beeswarm               0.4.0     2021-06-01 [1] CRAN (R 4.4.0)
   benchmarkme            1.0.8     2022-06-12 [1] CRAN (R 4.4.0)
   benchmarkmeData        1.0.4     2020-04-23 [1] CRAN (R 4.4.0)
   Biobase              * 2.64.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocFileCache          2.12.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocGenerics         * 0.50.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocIO                 1.14.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocManager            1.30.23   2024-05-04 [1] CRAN (R 4.4.0)
   BiocNeighbors          1.22.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocParallel         * 1.38.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   BiocSingular           1.20.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocVersion            3.19.1    2024-04-22 [1] Bioconductor 3.19 (R 4.4.0)
   Biostrings           * 2.72.1    2024-06-02 [1] Bioconductor 3.19 (R 4.4.0)
   bit                    4.0.5     2022-11-15 [1] CRAN (R 4.4.0)
   bit64                  4.0.5     2020-08-30 [1] CRAN (R 4.4.0)
   bitops                 1.0-7     2021-04-24 [1] CRAN (R 4.4.0)
   blob                   1.2.4     2023-03-17 [1] CRAN (R 4.4.0)
   bslib                  0.7.0     2024-03-29 [1] CRAN (R 4.4.0)
   cachem                 1.1.0     2024-05-16 [1] CRAN (R 4.4.0)
 P cli                    3.6.3     2024-06-21 [2] CRAN (R 4.4.0)
   codetools              0.2-20    2024-03-31 [1] CRAN (R 4.4.1)
   colorout             * 1.3-0.2   2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   colorspace             2.1-0     2023-01-23 [1] CRAN (R 4.4.0)
   config                 0.3.2     2023-08-30 [1] CRAN (R 4.4.0)
   cowplot                1.1.3     2024-01-22 [1] CRAN (R 4.4.0)
   crayon                 1.5.3     2024-06-20 [1] CRAN (R 4.4.0)
   curl                   5.2.1     2024-03-01 [1] CRAN (R 4.4.0)
   data.table           * 1.15.4    2024-03-30 [2] CRAN (R 4.4.0)
   DBI                    1.2.3     2024-06-02 [1] CRAN (R 4.4.0)
   dbplyr                 2.5.0     2024-03-19 [1] CRAN (R 4.4.0)
   default              * 1.0.0     2017-08-07 [1] CRAN (R 4.4.0)
   DelayedArray           0.30.1    2024-05-07 [1] Bioconductor 3.19 (R 4.4.0)
   DelayedMatrixStats     1.26.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   digest                 0.6.36    2024-06-23 [1] CRAN (R 4.4.0)
   doParallel             1.0.17    2022-02-07 [1] CRAN (R 4.4.0)
   dotCall64              1.1-1     2023-11-28 [1] CRAN (R 4.4.0)
   dplyr                  1.1.4     2023-11-17 [1] CRAN (R 4.4.0)
   DT                     0.33      2024-04-04 [1] CRAN (R 4.4.0)
   edgeR                  4.2.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   evaluate               0.24.0    2024-06-10 [1] CRAN (R 4.4.0)
   ExperimentHub          2.12.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   fansi                  1.0.6     2023-12-08 [1] CRAN (R 4.4.0)
   fastmap                1.2.0     2024-05-15 [1] CRAN (R 4.4.0)
   fields                 16.2      2024-06-27 [1] CRAN (R 4.4.0)
   filelock               1.0.3     2023-12-11 [1] CRAN (R 4.4.0)
   foreach                1.5.2     2022-02-02 [1] CRAN (R 4.4.0)
   generics               0.1.3     2022-07-05 [1] CRAN (R 4.4.0)
   GenomeInfoDb         * 1.40.1    2024-05-24 [1] Bioconductor 3.19 (R 4.4.0)
   GenomeInfoDbData       1.2.12    2024-05-01 [1] Bioconductor
   GenomicAlignments      1.40.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   GenomicRanges        * 1.56.1    2024-06-12 [1] Bioconductor 3.19 (R 4.4.1)
   ggbeeswarm             0.7.2     2023-04-29 [1] CRAN (R 4.4.0)
   ggplot2              * 3.5.1     2024-04-23 [1] CRAN (R 4.4.0)
   ggrepel                0.9.5     2024-01-10 [1] CRAN (R 4.4.0)
   glue                   1.7.0     2024-01-09 [1] CRAN (R 4.4.0)
   golem                  0.4.1     2023-06-05 [1] CRAN (R 4.4.0)
   gridExtra            * 2.3       2017-09-09 [1] CRAN (R 4.4.0)
   gtable                 0.3.5     2024-04-22 [1] CRAN (R 4.4.0)
   harmony              * 1.2.0     2023-11-29 [1] CRAN (R 4.4.0)
   here                   1.0.1     2020-12-13 [1] CRAN (R 4.4.0)
   htmltools              0.5.8.1   2024-04-04 [1] CRAN (R 4.4.0)
   htmlwidgets            1.6.4     2023-12-06 [1] CRAN (R 4.4.0)
   httpuv                 1.6.15    2024-03-26 [1] CRAN (R 4.4.0)
   httr                   1.4.7     2023-08-15 [1] CRAN (R 4.4.0)
   igraph                 2.0.3     2024-03-13 [1] CRAN (R 4.4.0)
   IRanges              * 2.38.1    2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   irlba                  2.3.5.1   2022-10-03 [1] CRAN (R 4.4.0)
   iterators              1.0.14    2022-02-05 [1] CRAN (R 4.4.0)
   jquerylib              0.1.4     2021-04-26 [1] CRAN (R 4.4.0)
   jsonlite               1.8.8     2023-12-04 [1] CRAN (R 4.4.0)
   KEGGREST               1.44.1    2024-06-19 [1] Bioconductor 3.19 (R 4.4.0)
   knitr                  1.48      2024-07-07 [1] CRAN (R 4.4.1)
   later                  1.3.2     2023-12-06 [1] CRAN (R 4.4.0)
   lattice                0.22-6    2024-03-20 [1] CRAN (R 4.4.1)
   lazyeval               0.2.2     2019-03-15 [1] CRAN (R 4.4.0)
   lifecycle              1.0.4     2023-11-07 [1] CRAN (R 4.4.0)
   limma                  3.60.3    2024-06-16 [1] Bioconductor 3.19 (R 4.4.0)
   locfit                 1.5-9.10  2024-06-24 [1] CRAN (R 4.4.0)
   magick                 2.8.3     2024-02-18 [1] CRAN (R 4.4.0)
   magrittr               2.0.3     2022-03-30 [1] CRAN (R 4.4.0)
   maps                   3.4.2     2023-12-15 [1] CRAN (R 4.4.0)
   Matrix                 1.7-0     2024-04-26 [1] CRAN (R 4.4.1)
   MatrixGenerics       * 1.16.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   matrixStats          * 1.3.0     2024-04-11 [1] CRAN (R 4.4.0)
   memoise                2.0.1     2021-11-26 [1] CRAN (R 4.4.0)
   mime                   0.12      2021-09-28 [1] CRAN (R 4.4.0)
   munsell                0.5.1     2024-04-01 [1] CRAN (R 4.4.0)
   paletteer              1.6.0     2024-01-21 [1] CRAN (R 4.4.0)
   pillar                 1.9.0     2023-03-22 [1] CRAN (R 4.4.0)
   pkgconfig              2.0.3     2019-09-22 [1] CRAN (R 4.4.0)
   plotly                 4.10.4    2024-01-13 [1] CRAN (R 4.4.0)
   png                    0.1-8     2022-11-29 [1] CRAN (R 4.4.0)
   promises               1.3.0     2024-04-05 [1] CRAN (R 4.4.0)
   purrr                  1.0.2     2023-08-10 [1] CRAN (R 4.4.0)
   R6                     2.5.1     2021-08-19 [1] CRAN (R 4.4.0)
   rappdirs               0.3.3     2021-01-31 [1] CRAN (R 4.4.0)
   RColorBrewer           1.1-3     2022-04-03 [1] CRAN (R 4.4.0)
   Rcpp                 * 1.0.13    2024-07-17 [1] CRAN (R 4.4.1)
   RcppAnnoy              0.0.22    2024-01-23 [1] CRAN (R 4.4.0)
   RCurl                  1.98-1.14 2024-01-09 [1] CRAN (R 4.4.0)
   rematch2               2.1.2     2020-05-01 [1] CRAN (R 4.4.0)
   ResidualMatrix         1.14.1    2024-06-23 [1] Bioconductor 3.19 (R 4.4.0)
   restfulr               0.0.15    2022-06-16 [1] CRAN (R 4.4.0)
   RhpcBLASctl            0.23-42   2023-02-11 [1] CRAN (R 4.4.0)
   rjson                  0.2.21    2022-01-09 [1] CRAN (R 4.4.0)
 P rlang                * 1.1.4     2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown              2.27      2024-05-17 [1] CRAN (R 4.4.0)
   rprojroot              2.0.4     2023-11-05 [1] CRAN (R 4.4.0)
   Rsamtools              2.20.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   RSQLite                2.3.7     2024-05-27 [1] CRAN (R 4.4.0)
   rstudioapi             0.16.0    2024-03-24 [1] CRAN (R 4.4.0)
   rsvd                   1.0.5     2021-04-16 [1] CRAN (R 4.4.0)
   rtracklayer            1.64.0    2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   S4Arrays               1.4.1     2024-05-20 [1] Bioconductor 3.19 (R 4.4.0)
   S4Vectors            * 0.42.1    2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   sass                   0.4.9     2024-03-15 [1] CRAN (R 4.4.0)
   ScaledMatrix           1.12.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scales                 1.3.0     2023-11-28 [1] CRAN (R 4.4.0)
   scater                 1.32.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scuttle                1.14.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   sessioninfo            1.2.2     2021-12-06 [1] CRAN (R 4.4.0)
   shiny                  1.8.1.1   2024-04-02 [1] CRAN (R 4.4.0)
   shinyWidgets           0.8.6     2024-04-24 [1] CRAN (R 4.4.0)
   SingleCellExperiment * 1.26.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   spam                   2.10-0    2023-10-23 [1] CRAN (R 4.4.0)
   SparseArray            1.4.8     2024-05-30 [1] Bioconductor 3.19 (R 4.4.0)
   sparseMatrixStats      1.16.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   SpatialExperiment    * 1.14.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   spatialLIBD          * 1.16.2    2024-05-28 [1] Bioconductor 3.19 (R 4.4.0)
   statmod                1.5.0     2023-01-06 [1] CRAN (R 4.4.0)
   SummarizedExperiment * 1.34.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   tibble                 3.2.1     2023-03-20 [1] CRAN (R 4.4.0)
   tidyr                  1.3.1     2024-01-24 [1] CRAN (R 4.4.0)
   tidyselect             1.2.1     2024-03-11 [1] CRAN (R 4.4.0)
   UCSC.utils             1.0.0     2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   utf8                   1.2.4     2023-10-22 [1] CRAN (R 4.4.0)
   uwot                   0.2.2     2024-04-21 [1] CRAN (R 4.4.0)
   vctrs                  0.6.5     2023-12-01 [1] CRAN (R 4.4.0)
   vipor                  0.4.7     2023-12-18 [1] CRAN (R 4.4.0)
   viridis                0.6.5     2024-01-29 [1] CRAN (R 4.4.0)
   viridisLite            0.4.2     2023-05-02 [1] CRAN (R 4.4.0)
   withr                  3.0.0     2024-01-16 [1] CRAN (R 4.4.0)
   xfun                   0.45      2024-06-16 [1] CRAN (R 4.4.0)
   XML                    3.99-0.17 2024-06-25 [1] CRAN (R 4.3.3)
   xtable                 1.8-4     2019-04-21 [1] CRAN (R 4.4.0)
   XVector              * 0.44.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   yaml                   2.3.9     2024-07-05 [1] CRAN (R 4.4.0)
   zlibbioc               1.50.0    2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

─────────────────────────────────────────────────────────────────────────────────
