==> spatial_HYP/code/04-Feature Selection, Harmony, MNN/03-Concat Feature Sets as One Supp Table.Rmd <==
---
title: "03-Concat Feature Sets as One Supp Table"
author: "Bernard Mulvey"
date: "2024-07-20"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
```

```{r}
feats <- readRDS("processed-data/04-feature_selection/01-hyp2n10_HVG10_20pctile_sample-blocked_nnSVG-mnrank-in-nomsigsamps_top10_20pctile.RDS")
genes <- unique(unlist(feats))
genes <- as.data.table(genes)

# get symbols
enslut <- fread("raw-data/USEFOR10X_ensg_lut_hg38_fromEns98_111123.txt")
enslut <- unique(enslut[,.(ensembl_gene_id,external_gene_name,hgnc_symbol)])

genes <- merge.data.table(genes,enslut,by.x="genes",by.y="ensembl_gene_id")
genes[,hvg10:=ifelse(genes %in% feats[["hyp2.hvg10"]],TRUE,FALSE)]
genes[,hvg20:=ifelse(genes %in% feats[["hyp2.hvg20"]],TRUE,FALSE)]
genes[,svg10:=ifelse(genes %in% feats[["nnsvg10"]],TRUE,FALSE)]
genes[,svg20:=ifelse(genes %in% feats[["nnsvg20"]],TRUE,FALSE)]

fwrite(genes,"local_supptab_assembly/Supp X-Feature sets.txt",sep='\t',quote=F)
```

### get the raw nnSVG results
```{r}
svgres <- readRDS("processed-data/04-feature_selection/01-hyp2n10_nnSVG_samplewise_n3s0.5.RDS")

i<-1
for (i in c(1:length(svgres))){
    tmp <- svgres[[i]]
    getnames <- c("gene_id","gene_name",grep(names(tmp),pattern="^nnSVG",value=T))
    tmp <- tmp[,..getnames]
    
    tmp[,sample:=names(svgres)[i]]

    if (i==1){
        svgres2 <- tmp
    } else {
        svgres2 <- rbind(svgres2,tmp)
    }
}

reord <- c("sample","gene_name",grep(names(svgres2),pattern="^nnSVG",value=T),"gene_id")
svgres2 <- svgres2[,..reord]

fwrite(svgres2,"local_supptab_assembly/Supp X-nnSVG raw.txt",sep='\t',quote=F)
```
